requirements.txt
streamlit
pandas
plotly
numpy
vnstock
PyPortfolioOpt
scipy
pandas-ta
statsmodels
google-generativeai
